+++ src/cli/formatter.ts
@@ -2,20 +2,21 @@ import fs from "fs";
 import { Formatter } from "../formatters";
 import { Report, Result } from "../reporter";
 
-type WrappedFormatter = (results: Result[]) => void;
+type WrappedFormatter = (results: Result[]) => string;
 
 function wrap(formatter: Formatter, dst: string) {
 	return (results: Result[]) => {
 		const output = formatter(results);
 		if (dst) {
 			fs.writeFileSync(dst, output, "utf-8");
+			return null;
 		} else {
-			process.stdout.write(output);
+			return output;
 		}
 	};
 }
 
-export function getFormatter(formatters: string): (report: Report) => void {
+export function getFormatter(formatters: string): (report: Report) => string {
 	const fn: WrappedFormatter[] = formatters.split(",").map(cur => {
 		const [name, dst] = cur.split("=", 2);
 		const moduleName = name.replace(/[^a-z]+/g, "");
@@ -23,6 +24,9 @@ export function getFormatter(formatters: string): (report: Report) => void {
 		return wrap(formatter, dst);
 	});
 	return (report: Report) => {
-		fn.forEach((formatter: WrappedFormatter) => formatter(report.results));
+		return fn
+			.map((formatter: WrappedFormatter) => formatter(report.results))
+			.filter(lines => lines && lines.length > 0)
+			.join("\n");
 	};
 }