+++ src/formatters/stylish.ts
@@ -1,4 +1,7 @@
 import { FormatterModule } from ".";
 
+const stylish = require("eslint/lib/formatters/stylish");
+export default stylish;
+
 declare const module: FormatterModule;
-module.exports = require("eslint/lib/formatters/stylish");
+module.exports = stylish;